pyproject.toml
[tool.poetry]
name = "g-drive-webhook"
version = "0.1.0"
description = ""
authors = ["Mikhail <mi.sidoren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.149.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "uvicorn app.app:app --reload"
